Control.Monad.LPMonad.Internal

Monad definitions

type LPM v c = LPT v c Identity

A simple monad for constructing linear programs. This library is intended to be able to link to a variety of different linear programming implementations.

type LPT v c = StateT (LP v c)

A simple monad transformer for constructing linear programs in an arbitrary monad.

runLPM :: (Ord v, Group c) => LPM v c a -> (a, LP v c)

runLPT :: (Ord v, Group c) => LPT v c m a -> m (a, LP v c)

execLPM :: (Ord v, Group c) => LPM v c a -> LP v c

Constructs a linear programming problem.

execLPT :: (Ord v, Group c, Monad m) => LPT v c m a -> m (LP v c)

Constructs a linear programming problem in the specified monad.

evalLPM :: (Ord v, Group c) => LPM v c a -> a

Runs the specified operation in the linear programming monad.

evalLPT :: (Ord v, Group c, Monad m) => LPT v c m a -> m a

Runs the specified operation in the linear programming monad transformer.

Constructing the LP

Objective configuration

setDirection :: MonadState (LP v c) m => Direction -> m ()

Sets the optimization direction of the linear program: maximization or minimization.

setObjective :: MonadState (LP v c) m => LinFunc v c -> m ()

Sets the objective function, overwriting the previous objective function.

addObjective :: (Ord v, Group c, MonadState (LP v c) m) => LinFunc v c -> m ()

Adds this function to the objective function.

addWeightedObjective :: (Ord v, Ring c, MonadState (LP v c) m) => c -> LinFunc v c -> m ()

Adds this function to the objective function, with the specified weight. Equivalent to addObjective (wt *^ obj).

Two-function constraints

leq :: (Ord v, Group c, MonadState (LP v c) m) => LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables. So, for example,

equal (f ^+^ g) h

constrains the value of h to be equal to the value of f plus the value of g.

equal :: (Ord v, Group c, MonadState (LP v c) m) => LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables. So, for example,

equal (f ^+^ g) h

constrains the value of h to be equal to the value of f plus the value of g.

geq :: (Ord v, Group c, MonadState (LP v c) m) => LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables. So, for example,

equal (f ^+^ g) h

constrains the value of h to be equal to the value of f plus the value of g.

leq' :: (Ord v, Group c, MonadState (LP v c) m) => String -> LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables, with a label on the constraint.

equal' :: (Ord v, Group c, MonadState (LP v c) m) => String -> LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables, with a label on the constraint.

geq' :: (Ord v, Group c, MonadState (LP v c) m) => String -> LinFunc v c -> LinFunc v c -> m ()

Specifies the relationship between two functions in the variables, with a label on the constraint.

One-function constraints

leqTo :: MonadState (LP v c) m => LinFunc v c -> c -> m ()

Sets a constraint on a linear function in the variables.

equalTo :: MonadState (LP v c) m => LinFunc v c -> c -> m ()

Sets a constraint on a linear function in the variables.

geqTo :: MonadState (LP v c) m => LinFunc v c -> c -> m ()

Sets a constraint on a linear function in the variables.

constrain :: MonadState (LP v c) m => LinFunc v c -> Bounds c -> m ()

The most general form of an unlabeled constraint.

leqTo' :: MonadState (LP v c) m => String -> LinFunc v c -> c -> m ()

Sets a labeled constraint on a linear function in the variables.

equalTo' :: MonadState (LP v c) m => String -> LinFunc v c -> c -> m ()

Sets a labeled constraint on a linear function in the variables.

geqTo' :: MonadState (LP v c) m => String -> LinFunc v c -> c -> m ()

Sets a labeled constraint on a linear function in the variables.

constrain' :: MonadState (LP v c) m => String -> LinFunc v c -> Bounds c -> m ()

The most general form of a labeled constraint.

Variable constraints

varLeq :: (Ord v, Ord c, MonadState (LP v c) m) => v -> c -> m ()

Sets a constraint on the value of a variable. If you constrain a variable more than once, the constraints will be combined. If the constraints are mutually contradictory, an error will be generated. This is more efficient than adding an equivalent function constraint.

varEq :: (Ord v, Ord c, MonadState (LP v c) m) => v -> c -> m ()

Sets a constraint on the value of a variable. If you constrain a variable more than once, the constraints will be combined. If the constraints are mutually contradictory, an error will be generated. This is more efficient than adding an equivalent function constraint.

varGeq :: (Ord v, Ord c, MonadState (LP v c) m) => v -> c -> m ()

Sets a constraint on the value of a variable. If you constrain a variable more than once, the constraints will be combined. If the constraints are mutually contradictory, an error will be generated. This is more efficient than adding an equivalent function constraint.

varBds :: (Ord v, Ord c, MonadState (LP v c) m) => v -> c -> c -> m ()

Bounds the value of a variable on both sides. If you constrain a variable more than once, the constraints will be combined. If the constraints are mutually contradictory, an error will be generated. This is more efficient than adding an equivalent function constraint.

setVarBounds :: (Ord v, Ord c, MonadState (LP v c) m) => v -> Bounds c -> m ()

The most general way to set constraints on a variable. If you constrain a variable more than once, the constraints will be combined. If you combine mutually contradictory constraints, an error will be generated. This is more efficient than creating an equivalent function constraint.

setVarKind :: (Ord v, MonadState (LP v c) m) => v -> VarKind -> m ()

Sets the kind ('type') of a variable. See VarKind.