Data.LinearProgram.GLPK.IO

Description

Bindings to the file I/O functions from GLPK, on the CPLEX LP file format.

Documentation

readLP :: (Ord v, Read v, Fractional c) => FilePath -> IO (LP v c)

Read a linear program from a file in CPLEX LP format. Warning: this will not necessarily succeed on all files generated by writeLP, as variable names may be changed.

readLP' :: FilePath -> IO (LP String Double)

Read a linear program from a file in CPLEX LP format.

writeLP :: (Ord v, Show v, Real c) => FilePath -> LP v c -> IO ()

Write a linear program to a file in CPLEX LP format.