Export

Description

Documentation

type Ex = CString -> IO CString

All FFI exported names in this module must have this opaque type , must be followed by Export, and must not use reserved symbols like "stdout" or "stdin".