FFI.TypeUncurry.DataKinds

Description

MessagePack FFI code adapted from call-haskell-from-anything

Documentation

data TypeList l where

Type-level list that can contain arbitrarily mixed types.

Example:

1 ::: "hello" ::: 2.3 :: TypeList '[Int, String, Double]

Constructors

Nil :: TypeList '[]

(:::) :: a -> TypeList l -> TypeList (a ': l) infixr 9

class ParamLength (l :: [*]) where

Allows to calculate the length of a TypeList, at compile time.

We need to use a Proxy for this.

Methods

paramLength :: Proxy l -> Int

Calculates the length of a type list, put into a proxy. Usage:

paramLength (Proxy :: Proxy l)

class (Param f ~ l, Result f ~ r) => ToTypeList f l r where

Function f can be translated to TypeList l with result type r.

Methods

translate :: f -> TypeList l -> r

Translates a function taking multiple arguments to a function taking a single TypeList containing the types of all arguments.

Example: t1 -> ... -> tn -> r becomes TypeList [t1, ..., tn] -> r.