FFI.TypeUncurry.Msgpack

Description

MessagePack FFI code adapted from call-haskell-from-anything

Documentation

class MessagePackRec l where

Helper to allow writing a MessagePack instance for TypeLists.

We need this because we have to call parseArray at the top-level MessagePack instance, but not at each function argument step.

Methods

fromObjectRec :: Monad m => [Object] -> m (TypeList l)

getTypeListFromMsgpackArray :: forall m l. (MessagePackRec l, ParamLength l, Monad m) => Object -> m (TypeList l)

Parses a tuple of arbitrary size (TypeLists) from a MessagePack array.

byteStringToCStringFunIO :: (ByteString -> IO ByteString) -> CString -> IO CString

Transforms a ByteString-mapping IO function to CString-mapping function for use in the FFI.

exportIO :: (MessagePack (TypeList l), ToTypeList f l (IO r), MessagePack r) => f -> CString -> IO CString

Orphan instances

(MessagePackRec l, ParamLength l) => MessagePack (TypeList l)

Instance details

Methods

toObject :: TypeList l -> Object

fromObject :: (Applicative m, Monad m) => Object -> m (TypeList l)