HBandit.BwCR

Description

Bandits with convex knapsacks and concave rewards (BwCR). See [1]

The parameters from the paper are: Convex set \(S \in [0,1]^d\). Convex loss function \(f : [0,1]^d -> [0,1]\) (written as concave objective in the paper). \(m\) arms.

Here, we implement the special LP case. This means that the case of \(k\) linearly-combined rewards (with weights \(\lambda\)) and \(c\) constraints (so that \(d = c+k\)) is implemented. More precisely, \(f(X) = \sum_{i=1..k} \lambda_i X_i\) and \(S = [0,1]^k \times [a_1,b_1] \times \ldots \times [a_c,b_c]\).

In this restricted case, the minimization(maximization in the original paper) inside the bandit iteration is a LP with k variables and c constraints. This LP is solved via GLPK bindings.

[1] Bandits with Global Convex Constraints and Objective, Shipra Agrawal, Nikhil R.Devanur. https://pubsonline.informs.org/doi/abs/10.1287/opre.2019.1840

Documentation

data BwCR a l

The BwCR state

Constructors

Screening (ScreeningBwCR a l)
Still screening for initial estimates

UCB (UCBBwCR a l)
The UCB sampling procedure has started.

data BwCRHyper a l

the BwCR hyperparameter \(\gamma\) used in \(rad_{\gamma(\nu,N)}\) used in the original paper.

Constructors

BwCRHyper

Fields

gamma :: Double

arms :: Arms a

constraints :: Interval l

objectiveWeights :: l

data ScreeningBwCR a l

Constructors

ScreeningBwCR

Fields

tScreening :: Int

screening :: a

screened :: [([ZeroOne Double], a)]

screenQueue :: [a]

screeningHyper :: BwCRHyper a l

data UCBBwCR a l

Constructors

UCBBwCR

Fields

t :: Int

lastAction :: a

k :: Int

weights :: NonEmpty (Weight a)

ucbHyper :: BwCRHyper a l

data Weight a

The information maintaining structure for one action.

Constructors

Weight

Fields

cumulativeEstimates :: [ZeroOne Double]

hits :: Int

action :: a

newtype T

Constructors

T Double

mkHyper :: Double -> T -> Arms a -> Interval [ZeroOne Double] -> [ZeroOne Double] -> BwCRHyper a [ZeroOne Double]

mkHyper \(\delta\) builds the hyperparameter for the original paper bound with confidence interval in probability \(1-\delta\).