HBandit.EpsGreedy

Description

This module implements the fixed rate \(\Epsilon\)-Greedy MAB algorithm.

The E(Epsilon)-Greedy algorithm selects a random action with probability \(\Epsilon\), or select the action with best average with probability \(1-\Epsilon\).

Documentation

data EpsGreedy a

The EpsGreedy state

Constructors

Screening (ScreeningGreedy a)
Still screening for initial estimates

ExploreExploit (ExploreExploitGreedy a)
The sampling procedure has started.

data Weight a

The information maintaining structure for one action.

Constructors

Weight

Fields

averageLoss :: Double

hits :: Int

action :: a

data EpsGreedyHyper a

The epsilon-greedy hyperparameter.

Constructors

EpsGreedyHyper

Fields

epsilon :: Double

arms :: Arms a

data ScreeningGreedy a

Constructors

ScreeningGreedy

Fields

tScreening :: Int

epsScreening :: Double

screening :: a

screened :: [(Double, a)]

screenQueue :: [a]

data ExploreExploitGreedy a

Constructors

ExploreExploitGreedy

Fields

t :: Int

eps :: Double

lastAction :: a

k :: Int

weights :: NonEmpty (Weight a)

pickreturn :: (RandomGen g, MonadState (EpsGreedy b) m) => ExploreExploitGreedy b -> g -> m (b, g)

Action selection and return

pickAction :: RandomGen g => ExploreExploitGreedy a -> g -> (a, g)

Action selection primitive

updateAvgLoss :: Double -> Weight a -> Weight a

This is a rudimentary online mean accumulator. Its numerical properties are probably bad, but we don't need much more here.