HBandit.Types

Description

Utility functions for MAB algorithms.

Documentation

type ZeroOne l = Refined (FromTo 0 1) l

Type alias for a \(\mathbb{L}=[0,1]\) refined value.

zero :: (Ord a, Num a) => ZeroOne a

0

one :: (Ord a, Num a) => ZeroOne a

1

normalize :: Double -> Double -> Maybe (ZeroOne Double)

normalize x xm normalizes x using xm. Returns Nothing in case the resulting value is not contained in $[0,1]$ .

normalizedSum :: (Ord a, Num a) => [(a, ZeroOne a)] -> ZeroOne a

normalizedSum sums the fst from the tuple list using weights in thesnd. The result is normalized against the list of snds.