HBandit.Util

Description

Utility functions for MAB algorithms.

Documentation

sampleWL :: RandomGen g => NonEmpty (ZeroOne Double, a) -> g -> (a, g)

Samples from a weighted probability distribution. The sum of weights must not be zero.

normalize :: (Floating a, Ord a) => a -> a -> Maybe (ZeroOne a)

normalize x xm normalizes x using xm. Returns Nothing in case the resulting value is not contained in $[0,1]$ .

unsafeNormalizePanic :: (Floating a, Ord a) => a -> a -> ZeroOne a

normalize x xm normalizes x using xm. Returns Nothing in case the resulting value is not contained in $[0,1]$ .

normalizeDistribution :: (Floating p, Ord p) => NonEmpty (p, a) -> Maybe (NonEmpty (ZeroOne p, a))

normalizedSum :: (Ord a, Fractional a) => [(ZeroOne a, ZeroOne a)] -> ZeroOne a

normalizedSum sums the snd from the tuple list using weights in the fst. The result is normalized against the list of fsts.