NRM.Behavior

Description

external interface

behavior :: Cfg -> NRMState -> Time -> NRMEvent -> IO (NRMState, [Behavior])

External interface for NRM behavior, without RWS monad

internal dispatching

nrm :: Time -> NRMEvent -> NRM ()

The nrm function contains the main logic of the NRM daemon. It changes the state and produces an associated behavior to be executed by the runtime. This contains the slice management logic, the sensor callback logic, the control loop callback logic. Works in the @NRM monad.

nrmDownstreamEvent :: Time -> DownstreamClientID -> Event -> NRM (CommonOutcome Measurement)

Downstream event handler.

doControl :: Input -> NRM ()

doControl checks the integrator state and triggers a control iteration if NRM is ready.

commonSP :: Time -> ActiveSensorKey -> Double -> NRM (CommonOutcome Measurement)

commonProcess :: Time -> MeasurementOutput -> NRM (CommonOutcome Measurement)

folder :: Time -> (NRMState, Maybe [Measurement]) -> (PassiveSensorKey, ScopedLens NRMState PassiveSensor) -> IO (NRMState, Maybe [Measurement])

mayInjectLibnrmPreload :: Cfg -> Manifest -> Env -> Env

mayRep :: Cmd -> Rep -> Behavior

respondContent :: Text -> Cmd -> CmdID -> OutputType -> Behavior