NRM.CPD

Description

This module is responsible for runtime generation of the concrete CPD description to be optimized by the control loop.

Documentation

toCPD :: Maybe ControlCfg -> NRMState -> Problem

throughputConstrained

Arguments

:: ControlCfg
Control configuration

-> NRMState
State

-> ([(Double, OExpr)], [(Double, OExpr)])

addAll :: NonEmpty SensorID -> OExpr

produces a sum objective normalized by #sensors

maybeMinus :: Maybe OExpr -> Maybe OExpr -> Maybe OExpr

Subtract two objectives, defaulting to either of them if one is absent.