NRM.State

Description

Initial state

initialState :: Cfg -> Time -> IO NRMState

Populate the initial NRMState.

Creation/Registration

createSlice :: SliceID -> NRMState -> NRMState

Registers a slice if not already tracked in the state, and returns the new state.

registerAwaiting :: CmdID -> CmdCore -> SliceID -> NRMState -> NRMState

Registers an awaiting command in an existing slice

registerFailed :: CmdID -> NRMState -> Maybe (NRMState, SliceID, Slice, CmdCore)

Fails an awaiting command.

registerLaunched :: CmdID -> ProcessID -> NRMState -> Either Text (NRMState, SliceID, Maybe UpstreamClientID)

Turns an awaiting command to a launched one.

Removal

Command removal

data CmdKey

Wrapper for the type of key to lookup commands on

Constructors

KCmdID CmdID

KProcessID ProcessID

data DeletionInfo

Result annotation for command removal from the state.

Constructors

SliceRemoved
If the slice was removed as a result of the command deletion

CmdRemoved
If the command was removed but the slice stayed.

removeCmd :: CmdKey -> NRMState -> Maybe (DeletionInfo, CmdID, Cmd, SliceID, NRMState)

Removes a command from the state, and also removes the slice if it's empty as a result.