NRM.Types.Behavior

Description

Documentation

data NRMEvent

The Behavior datatype describes an event from the runtime on which to react.

Constructors

Req UpstreamClientID Req
A Request was received on the Upstream API.

RegisterCmd CmdID CmdStatus
Registering a child process.

DownstreamEvent DownstreamClientID Event
Event from the application side.

DoOutput CmdID OutputType Text
Stdin/stdout data from the app side.

ChildDied ProcessID ExitCode
Child death event

DoSensor
Sensor callback

DoControl
Control loop calback

DoShutdown
Shutting down the daemon

The Behavior specification

data Behavior

The Behavior datatype encodes a behavior to be executed by the NRM runtime.

Constructors

Log DaemonVerbosity Text
Log a message

Rep UpstreamClientID Rep
Reply to an upstream client.

Pub Pub
Publish messages on upstream

StartChild CmdID Command Arguments Env
Start a child process

KillChildren [CmdID] [(UpstreamClientID, Rep)]
Kill children processes and send some messages back upstream.

ClearChild CmdID (Maybe (UpstreamClientID, Rep))
Pop one child process and may send a message back upstream.

data CmdStatus

The Launch status of a command, for registration.

Constructors

Launched ProcessID
In case the command to start a child succeeded, mark it as registered and provide its PID.

NotLaunched
In case the command to start a child failed.