NRM.Types.MemBuffer

Description

Documentation

type MemBuffer a = Refined (SizeLessThan 10) [a]

Memory buffer with 10 points.

empty :: MemBuffer a

empty MemBuffer

enqueue :: a -> MemBuffer a -> MemBuffer a

adding a data point to a MemBuffer, discarding the oldest point if there are more than 10 points.

avgBuffer :: Floating a => MemBuffer a -> a

adding a data point to a MemBuffer, discarding the oldest point if there are more than 10 points.