NRM.Types.Messaging.UpstreamReq

Description

Documentation

data Req

Constructors

ReqSliceList SliceList
encodes a request for the list of current slices

ReqRun Run
encodes a request to run a command

ReqKillSlice KillSlice
encodes a request to kill a whole slice

ReqCPD CPD
encodes a request to obtain the current control problem description

ReqKillCmd KillCmd
encodes a request to kill a running command

ReqSetPower SetPower
encodes a request to set the global power bound

ReqGetState GetState
encodes a request to obtain the full daemon state

ReqGetConfig GetConfig
encodes a request to get the daemon configuration

data Run

Constructors

Run

Fields

manifest :: Manifest

spec :: CmdSpec

runSliceID :: SliceID

detachCmd :: Bool

data SliceList

Constructors

SliceList

data CPD

Constructors

CPD

newtype KillSlice

Constructors

KillSlice

Fields

killSliceID :: SliceID

newtype KillCmd

Constructors

KillCmd

Fields

killCmdID :: CmdID

newtype SetPower

Constructors

SetPower

Fields

limit :: Power

data GetConfig

Constructors

GetConfig

data GetState

Constructors

GetState