NRM.Types.Slice

Description

Documentation

data Slice

NRM's internal view of the state of a slice.

Constructors

Slice

Fields

cmds :: Map CmdID Cmd

map of running commands

awaiting :: Map CmdID CmdCore

map of commands awaiting to be registered as running by the runtime

emptySlice :: Slice

Constructor for an empty slice.

insertCmd :: CmdID -> Cmd -> Slice -> Slice

Insert a running command in a slice (with replace)

data SliceID

Constructors

SliceID UUID

Name Text

nextSliceID :: IO (Maybe SliceID)

parseSliceID :: Text -> SliceID

toText :: SliceID -> Text